% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{get_edgeR_res}
\alias{get_edgeR_res}
\title{Get edgeR Results}
\usage{
get_edgeR_res(
  se,
  contrasts,
  design = NULL,
  res.list = list(),
  block = NULL,
  lfc.th = NULL,
  norm.ercc = FALSE,
  ercc.pattern = "^ERCC-",
  use.lrt = FALSE
)
}
\arguments{
\item{se}{A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} or \link[edgeR:DGEList]{edgeR::DGEList} object.}

\item{contrasts}{A named list of contrasts, e.g. \code{list("condition_AvB" = c("condition", "A", "B"))}.
The first element is the variable of interest, the second is the test, and the third is the reference level.
The name of  each element in the list will be used as a name in the results table.}

\item{design}{The design matrix, as output from \code{model.matrix()}.
If not provided, the function will create a design matrix based on the contrasts and \code{block} (if provided).}

\item{res.list}{A named list to hold edgeR results.
Allows the function to be run multiple times if needed and append to the same list.
Defaults to an empty list.}

\item{block}{A vector of additional terms to be considered in the model, beyond the main effect.
Defaults to \code{NULL}.}

\item{lfc.th}{A numeric vector of log2 fold-change thresholds for testing.
The function will run \code{glmTreat()} for each threshold specified.
Defaults to \code{NULL}.}

\item{norm.ercc}{A logical indicating whether to normalize to ERCC spike-ins.}

\item{ercc.pattern}{A character string indicating the pattern to match ERCC spike-ins.
Defaults to \verb{^ERCC-}.}

\item{use.lrt}{A logical indicating whether to use the likelihood ratio test (LRT)
instead of the quasi-likelihood F-test.}
}
\value{
A named list of \link[edgeR:topTags]{edgeR::TopTags} objects for the specified contrasts.
}
\description{
This function obtains a set of comparisons from a edgeR analysis, given a named list of contrasts.
It allows additional model
parameters to be specified and a design matrix to be manually adjusted.
}
\details{
This function is designed to work with relatively simple designs,
e.g. variable of interest with a blocking factor or two.
If you need a bunch of interaction terms or nested designs, this function is probably not for you.
}
\examples{
nlibs <- 4
ngenes <- 1000
dispersion.true <- 1/rchisq(ngenes, df=10)
group <- factor(c("A", "A", "B", "B"))
design <- model.matrix(~group)

# Generate count data
y <- rnbinom(ngenes*nlibs,mu=20,size=1/dispersion.true)
y <- matrix(y,ngenes,nlibs)
d <- DGEList(y)

res <- get_edgeR_res(d, contrasts = list("condition_AvB" = c("group", "A", "B")))
names(res)
head(res[[1]])

}
\author{
Jared Andrews
}
