% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{GSEA_barplot}
\alias{GSEA_barplot}
\title{Plot top Gene Set Enrichment Analysis (GSEA) results in a barplot.}
\usage{
GSEA_barplot(
  gsea.res,
  genesets.name = NULL,
  padj.th = 0.05,
  top = 35,
  color.by.direction = FALSE,
  up.color = "red",
  down.color = "blue"
)
}
\arguments{
\item{gsea.res}{A data.frame of GSEA results as returned by \code{\link[fgsea:fgsea]{fgsea::fgsea()}}.}

\item{genesets.name}{A name for the gene set collection or group, used to label plot.}

\item{padj.th}{The significance threshold (adjusted p-value) for filtering gene sets
Defaults to 0.05.}

\item{top}{The number of top significant gene sets in each direction to consider.
Defaults to 35.}

\item{color.by.direction}{Logical. If \code{TRUE}, bars are colored by up/down direction instead of significance.
Defaults to FALSE.}

\item{up.color}{Color for upregulated pathways if \code{color.by.direction} is \code{TRUE}.
Defaults to "red".}

\item{down.color}{Color for downregulated pathways if \code{color.by.direction} is \code{TRUE}.
Defaults to "blue".}
}
\value{
A ggplot object.
}
\description{
This function condenses GSEA results by plotting the top significant gene sets.
}
\examples{
@examples
\dontrun{
GSEA_barplot(gsea.list = my_gsea_results, outdir = "./summary", padj.th = 0.01, top = 50)
}

}
\author{
Jared Andrews
}
