% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_enrichment2}
\alias{run_enrichment2}
\title{Run Enrichment Analysis}
\usage{
run_enrichment2(
  de.list = NULL,
  genes = NULL,
  bg = NULL,
  method = c("KEGG", "Reactome", "GO", "universal"),
  species = c("human", "mouse"),
  ont = c("BP", "MF", "CC", "ALL"),
  name = "sample",
  sig.th = 0.05,
  sig.col = "padj",
  lfc.th = 0,
  lfc.col = "log2FoldChange",
  OrgDb = "org.Hs.eg.db",
  id.col = "ENSEMBL",
  id.type = "ENSEMBL",
  ...
)
}
\arguments{
\item{de.list}{Named list of DE results data.frames for group enrichment.
If provided, \code{genes} and \code{bg} will be ignored.
Default \verb{is NULL}.}

\item{genes}{Character vector of gene IDs for simple enrichment.
Default \verb{is NULL}.}

\item{bg}{Character vector of gene IDs to be used as background for simple enrichment.
Default \verb{is NULL}.}

\item{method}{Enrichment method to use. Options are "KEGG", "Reactome", "GO", or "universal".}

\item{species}{Species to use. Options are "human" or "mouse".
This determines the organism values for KEGG and Reactome.}

\item{ont}{Character vector of GO ontologies to test, options include "BP", "MF", "CC", and "ALL".
Default is \code{c("BP", "MF", "CC", "ALL")}.}

\item{name}{Output prefix name for simple enrichment if \code{genes} and \code{bg} provided.}

\item{sig.th}{Significance threshold for DE genes.
Default is 0.05.}

\item{sig.col}{Name of the column in the results data.frame containing significance values to use.
Default is "padj".}

\item{lfc.th}{Log fold change threshold for DE genes.
Default is 0.}

\item{lfc.col}{Name of the column in the results data.frame containing log2 fold change values.
Default is "log2FoldChange".}

\item{OrgDb}{Annotation database to use (default: "org.Hs.eg.db").}

\item{id.col}{Name of gene ID column (default: "ENSEMBL").}

\item{id.type}{Type of gene ID used (default: "ENSEMBL").}

\item{...}{Additional arguments passed to the enrichment functions.}

\item{outdir}{Output directory (default: "./enrichments").}
}
\value{
Named list of enrichment results.
}
\description{
This function performs enrichment analysis using KEGG, Reactome, GO, or a custom universal method.
It accepts either a list of differential expression results for group enrichment or
a set of genes and background for simple enrichment.
}
\author{
Jared Andrews
}
