% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{run_GSEA}
\alias{run_GSEA}
\title{Run Gene Set Enrichment Analysis (GSEA) with arbitrary gene sets.}
\usage{
run_GSEA(
  sigs,
  ranked.genes,
  outdir,
  res.name = "GSEA",
  res.list = list(),
  padj.th = 0.05,
  min.size = 15,
  max.size = 1000,
  BPPARAM = NULL,
  ...
)
}
\arguments{
\item{sigs}{A named list of gene signatures.}

\item{ranked.genes}{A named list of ranked genes.
Using a list allows for results from many comparisons to be passed.}

\item{outdir}{The output directory for results.}

\item{res.name}{A name to be assigned to for the results added to \code{res.list}.
Also used as a prefix for output files.}

\item{res.list}{A list to store results. Useful for saving lots of results to a single file,
e.g. from multiple comparisons.
Defaults to an empty list.}

\item{padj.th}{The adjusted p-value threshold for limiting which gene sets to plot.
Defaults to 0.05.}

\item{min.size}{The minimum size of the gene sets to be considered for GSEA.
Defaults to 15.}

\item{max.size}{The maximum size of the gene sets to be considered for GSEA.
Defaults to 1000.}

\item{BPPARAM}{The BiocParallelParam object specifying the parallel back-end to be used.
Defaults to NULL.}

\item{...}{Additional arguments to pass to the \code{\link[fgsea:fgsea]{fgsea::fgsea()}} function.}
}
\value{
A list of GSEA results if \code{res.list} is not `NULL``, otherwise, results are saved as files in the specified output directory.
}
\description{
This function performs GSEA on a named list of ranked genes.
}
\details{
The function creates various output files including detailed GSEA results, enrichment plots, and tables of top enriched pathways.
}
\examples{
\dontrun{
run_GSEA(
    sigs = msigdb, ranked.genes = my_genes, outdir = "./results", outprefix = "experiment1",
    res.list = list()
)
}

}
\author{
Jared Andrews
}
