% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{get_DESeq2_res}
\alias{get_DESeq2_res}
\title{Get DESeq2 Results}
\usage{
get_DESeq2_res(
  dds,
  res.list,
  contrasts,
  user.mat = FALSE,
  block = NULL,
  design = NULL,
  alpha = 0.05,
  lfc.th = c(log2(1.25), log2(1.5)),
  shrink.method = "ashr",
  norm.ercc = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{dds}{An object of class DESeqDataSet.}

\item{res.list}{A list of DESeq2 result tables.
Allows the function to be run multiple times if needed and append to the same list.}

\item{contrasts}{A named list of contrasts.}

\item{user.mat}{A logical indicating whether a user-specified model matrix is provided.
Defaults to FALSE.}

\item{block}{A vector of additional terms to be considered in the model, beyond the main effect.
Defaults to NULL.}

\item{design}{The design formula or matrix.
If a matrix is provided, ensure 'user.mat' is set to TRUE.
Defaults to NULL.}

\item{alpha}{The significance level for hypothesis testing.
Defaults to 0.05.}

\item{lfc.th}{A numeric vector of log2 fold-change thresholds.
Defaults to c(log2(1.25), log2(1.5)).}

\item{shrink.method}{The method used for shrinkage estimation.
Defaults to "ashr".}

\item{norm.ercc}{A logical indicating whether to normalize to ERCC spike-ins.}

\item{BPPARAM}{The BiocParallelParam object specifying the parallel back-end to be used.
Defaults to NULL.}
}
\value{
A named list of DESeq2 result tables for the specified contrasts.
}
\description{
This function obtains a set of comparisons from a DESeq2 analysis, given a named list of contrasts.
It allows additional model
parameters to be specified and a design matrix to be manually adjusted.
}
\examples{
\dontrun{
get_DESEQ2_res(dds, res.list, contrasts,
    user.mat = TRUE, block = c("term1", "term2"),
    design = my_design, alpha = 0.05, lfc.th = log2(2),
    shrink.method = "apeglm", outdir = "./my_results", BPPARAM = MulticoreParam(2)
)
}

}
\author{
Jared Andrews
}
