% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_enrichment}
\alias{run_enrichment}
\title{Run Enrichment Analysis}
\usage{
run_enrichment(
  genes,
  bg,
  res.name,
  TERM2GENE = NULL,
  res.list = list(),
  method = c("KEGG", "Reactome", "GO", "universal"),
  species = c("human", "mouse"),
  ont = c("BP", "MF", "CC", "ALL"),
  OrgDb = "org.Hs.eg.db",
  id.type = "ENSEMBL",
  ...
)
}
\arguments{
\item{genes}{Character vector of gene IDs for simple enrichment.
Default is \code{NULL}.}

\item{bg}{Character vector of gene IDs to be used as background for simple enrichment.
Default is \code{NULL}.}

\item{res.name}{Output prefix name for results.}

\item{TERM2GENE}{data.frame of two columns, the first for the term and the second for the gene ID.
Each gene in each geneset gets its own row (long format). Gene identifiers should be ENTREZID.
Required if \code{method} is "universal".}

\item{res.list}{Named list to which enrichment results should be added.
Default is an empty list.}

\item{method}{Enrichment method to use. Options are "KEGG", "Reactome", "GO", or "universal".}

\item{species}{Species to use. Options are "human" or "mouse".
This determines the organism values for KEGG and Reactome.}

\item{ont}{Character vector of GO ontologies to test, options include "BP", "MF", "CC", and "ALL".
Default is \code{c("BP", "MF", "CC", "ALL")}.}

\item{OrgDb}{Annotation database to use (default: "org.Hs.eg.db").}

\item{id.type}{Type of gene ID used (default: "ENSEMBL").}

\item{...}{Additional arguments passed to the enrichment functions.}
}
\value{
Named list of enrichment results.
}
\description{
This function performs enrichment analysis using KEGG, Reactome, GO, or a custom universal method.
It accepts a set of genes and background genes for enrichment.
}
\details{
This function will not return results for analyses with no significant results.
}
\author{
Jared Andrews
}
